revise according to database

+++ YuRead/server/src/main/java/com/lcl/demo/sbDemo/dao/NoteMapper.java
@@ -3,13 +3,13 @@ package com.lcl.demo.sbDemo.dao;
 import com.lcl.demo.sbDemo.entity.Note;
 
 public interface NoteMapper {
-    int deleteByPrimaryKey(Integer noteId);
+    int deleteByPrimaryKey(Long noteId);
 
     int insert(Note record);
 
     int insertSelective(Note record);
 
-    Note selectByPrimaryKey(Integer noteId);
+    Note selectByPrimaryKey(Long noteId);
 
     int updateByPrimaryKeySelective(Note record);
